refact: dependencies update and reorganisation
