+++ buildSrc/src/main/kotlin/dependency/android/AndroidArchCore.kt
@@ -1,0 +1,10 @@
+package dependency.android
+
+import dependency.Dependency
+
+object AndroidArchCore : Dependency {
+
+    override val group: String = "androidx.arch.core"
+    override val artifact: String = "core"
+    override val version: String = "2.1.0"
+}
