+++ buildSrc/src/main/kotlin/dependency/android/Paging.kt
@@ -1,4 +1,4 @@
-package dependency.ui
+package dependency.android
 
 import dependency.Dependency
