+++ buildSrc/src/main/kotlin/dependency/async/Coroutines.kt
@@ -1,4 +1,4 @@
-package dependency.network
+package dependency.async
 
 import dependency.Dependency
